requirements.txt
django
djangorestframework
jsonfield
coreapi
coreapi-cli
py-cpuinfo
psutil
gpustat